+++ index.html
@@ -232,8 +232,9 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="/js/functions.js"></script>
-  <script src="/js/main.js"></script>
+  <!--<script src="/js/functions.js"></script>-->
+  <script src="/js/main.js" type="module"></script>
+  <script src="/js/util.js" type="module"></script>
+  <script src="/js/data.js" type="module"></script>
 </body>
-
 </html>
